Further breakdown db lib into separate modules
