+++ web/src/db/error.rs
@@ -1,0 +1,17 @@
+#[derive(serde::Serialize, Debug, thiserror::Error)]
+pub enum DbError {
+    #[error("Unknown error")]
+    UnknownError,
+
+    #[error("Database error: {0}")]
+    Database(String),
+
+    #[error("RecordNotFound: {0}")]
+    RecordNotFound(String),
+}
+
+impl From<sqlx::Error> for DbError {
+    fn from(value: sqlx::Error) -> Self {
+        Self::Database(value.to_string())
+    }
+}
