+++ web/src/db/mod.rs
@@ -1,2 +1,4 @@
-pub mod lib;
+pub mod error;
+pub mod finders;
+pub mod pool;
 pub mod station_record;
